chore: rename type generators to inline converter generators
