+++ Core/NosSmooth.PacketSerializersGenerator/InlineConverterGenerators/BoolInlineConverterGenerator.cs
@@ -1,5 +1,5 @@
 //
-//  FallbackInlineConverterGenerator.cs
+//  BoolInlineConverterGenerator.cs
 //
 //  Copyright (c) František Boháček. All rights reserved.
 //  Licensed under the MIT license. See LICENSE file in the project root for full license information.
@@ -9,31 +9,34 @@ using NosSmooth.PacketSerializersGenerator.Data;
 using NosSmooth.PacketSerializersGenerator.Errors;
 using NosSmooth.PacketSerializersGenerator.Extensions;
 
-namespace NosSmooth.PacketSerializersGenerator.TypeGenerators;
+namespace NosSmooth.PacketSerializersGenerator.InlineConverterGenerators;
 
 /// <inheritdoc />
-public class FallbackInlineConverterGenerator : IInlineConverterGenerator
+public class BoolInlineConverterGenerator : IInlineConverterGenerator
 {
     /// <inheritdoc />
     public bool ShouldHandle(ParameterInfo parameter)
-    {
-        return true;
-    }
+        => parameter.Parameter.Type!.ToString() == "bool";
 
     /// <inheritdoc />
     public IError? GenerateSerializerPart(IndentedTextWriter textWriter, PacketInfo packet)
     {
         var parameter = packet.Parameters.Current;
-        textWriter.WriteMultiline
-        (
-            $@"
-var {parameter.GetResultVariableName()} = _typeConverterRepository.Serialize<{parameter.GetActualType()}>(obj.{parameter.Name}, builder);
-if (!{parameter.GetResultVariableName()}.IsSuccess)
-{{
-    return Result.FromError(new PacketParameterSerializerError(this, ""{parameter.Name}"", {parameter.GetResultVariableName()}), {parameter.GetResultVariableName()});
-}}
-"
-        );
+        if (parameter.Nullable)
+        {
+            textWriter.WriteLine($"if (obj.{parameter.Name} is null)");
+            textWriter.WriteLine("{");
+            textWriter.Indent++;
+            textWriter.WriteLine("builder.Append(\"-\");");
+            textWriter.Indent--;
+            textWriter.WriteLine("}");
+            textWriter.WriteLine("else");
+        }
+        textWriter.WriteLine("{");
+        textWriter.Indent++;
+        textWriter.WriteLine($"builder.Append(obj.{parameter.Name} ? \"1\" : \"0\");");
+        textWriter.Indent--;
+        textWriter.WriteLine("}");
         return null;
     }
 
@@ -43,13 +46,13 @@ if (!{parameter.GetResultVariableName()}.IsSuccess)
         var parameter = packet.Parameters.Current;
         string isLastString = packet.Parameters.IsLast ? "true" : "false";
         textWriter.WriteMultiline($@"
-var {parameter.GetResultVariableName()} = _typeConverterRepository.Deserialize<{parameter.GetNullableType()}>(stringEnumerator);
+var {parameter.GetResultVariableName()} = stringEnumerator.GetNextToken();
 var {parameter.GetErrorVariableName()} = CheckDeserializationResult({parameter.GetResultVariableName()}, ""{parameter.Name}"", stringEnumerator, {isLastString});
 if ({parameter.GetErrorVariableName()} is not null)
 {{
     return Result<{packet.Name}?>.FromError({parameter.GetErrorVariableName()}, {parameter.GetResultVariableName()});
 }}
-var {parameter.GetNullableVariableName()} = {parameter.GetResultVariableName()}.Entity;
+{parameter.GetNullableType()} {parameter.GetNullableVariableName()} = {parameter.GetResultVariableName()}.Entity.Token == ""1"" ? true : ({parameter.GetResultVariableName()}.Entity.Token == ""-"" ? null : false);
 ");
         return null;
     }
