+++ Core/NosSmooth.PacketSerializersGenerator/InlineConverterGenerators/EnumInlineConverterGenerator.cs
@@ -1,37 +1,31 @@
 //
-//  BasicTypeGenerator.cs
+//  EnumInlineConverterGenerator.cs
 //
 //  Copyright (c) František Boháček. All rights reserved.
 //  Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.CodeDom.Compiler;
+using Microsoft.CodeAnalysis;
 using NosSmooth.PacketSerializersGenerator.Data;
 using NosSmooth.PacketSerializersGenerator.Errors;
 using NosSmooth.PacketSerializersGenerator.Extensions;
 
-namespace NosSmooth.PacketSerializersGenerator.TypeGenerators;
+namespace NosSmooth.PacketSerializersGenerator.InlineConverterGenerators;
 
-/// <summary>
-/// Serializes and deserializes
-/// long, ulong, int, uint, short, ushort, byte, sbyte.
-/// </summary>
-public class BasicTypeGenerator : IInlineConverterGenerator
+/// <inheritdoc />
+public class EnumInlineConverterGenerator : IInlineConverterGenerator
 {
-    /// <summary>
-    /// The list of the types to handle.
-    /// </summary>
-    public static IReadOnlyList<string> HandleTypes => new[] { "long", "ulong", "int", "uint", "short", "ushort", "byte", "sbyte" };
-
     /// <inheritdoc />
     public bool ShouldHandle(ParameterInfo parameter)
-        => HandleTypes.Contains(parameter.Parameter.Type!.ToString());
+        => parameter.Type.TypeKind == TypeKind.Enum;
 
     /// <inheritdoc />
     public IError? GenerateSerializerPart(IndentedTextWriter textWriter, PacketInfo packet)
     {
         var parameter = packet.Parameters.Current;
+        var underlyingType = ((INamedTypeSymbol)parameter.Type).EnumUnderlyingType!.ToString();
         textWriter.WriteLine
-            ($"builder.Append(obj.{parameter.Name}{(parameter.Nullable ? "?" : string.Empty)}.ToString() ?? \"-\");");
+            ($"builder.Append((({underlyingType}?)obj.{parameter.Name}{(parameter.Nullable ? "?" : string.Empty)}).ToString() ?? \"-\");");
         return null;
     }
 
@@ -39,7 +33,7 @@ public class BasicTypeGenerator : IInlineConverterGenerator
     public IError? GenerateDeserializerPart(IndentedTextWriter textWriter, PacketInfo packet)
     {
         var parameter = packet.Parameters.Current;
-        var type = parameter.Parameter.Type!.ToString();
+        var underlyingType = ((INamedTypeSymbol)parameter.Type).EnumUnderlyingType!.ToString();
         string isLastString = packet.Parameters.IsLast ? "true" : "false";
         textWriter.WriteMultiline
         (
@@ -51,14 +45,15 @@ if ({parameter.GetErrorVariableName()} is not null)
     return Result<{packet.Name}?>.FromError({parameter.GetErrorVariableName()}, {parameter.GetResultVariableName()});
 }}
 {parameter.GetVariableName()} = default;
+{underlyingType} {parameter.GetVariableName()}Underlying = default;
 {parameter.GetNullableType()} {parameter.GetNullableVariableName()};
 if ({parameter.GetResultVariableName()}.Entity.Token == ""-"") {{
     {parameter.GetNullableVariableName()} = null;
 }}
-else if (!{type}.TryParse({parameter.GetResultVariableName()}.Entity.Token, out {parameter.GetVariableName()})) {{
-    return new PacketParameterSerializerError(this, ""{parameter.Name}"", {parameter.GetResultVariableName()}, $""Could not convert {{{parameter.GetResultVariableName()}.Entity.Token}} as {type} in inline converter"");
+else if (!{underlyingType}.TryParse({parameter.GetResultVariableName()}.Entity.Token, out {parameter.GetVariableName()}Underlying)) {{
+    return new PacketParameterSerializerError(this, ""{parameter.Name}"", {parameter.GetResultVariableName()}, $""Could not convert {{{parameter.GetResultVariableName()}.Entity.Token}} as {underlyingType} in inline converter"");
 }}
-{parameter.GetNullableVariableName()} = {parameter.GetVariableName()};
+{parameter.GetNullableVariableName()} = ({parameter.GetNullableType()}){parameter.GetVariableName()}Underlying;
 "
         );
         return null;
