+++ Core/NosSmooth.PacketSerializersGenerator/InlineConverterGenerators/StringInlineConverterGenerator.cs
@@ -1,5 +1,5 @@
 //
-//  BoolTypeGenerator.cs
+//  StringInlineConverterGenerator.cs
 //
 //  Copyright (c) František Boháček. All rights reserved.
 //  Licensed under the MIT license. See LICENSE file in the project root for full license information.
@@ -9,34 +9,20 @@ using NosSmooth.PacketSerializersGenerator.Data;
 using NosSmooth.PacketSerializersGenerator.Errors;
 using NosSmooth.PacketSerializersGenerator.Extensions;
 
-namespace NosSmooth.PacketSerializersGenerator.TypeGenerators;
+namespace NosSmooth.PacketSerializersGenerator.InlineConverterGenerators;
 
 /// <inheritdoc />
-public class BoolTypeGenerator : IInlineConverterGenerator
+public class StringInlineConverterGenerator : IInlineConverterGenerator
 {
     /// <inheritdoc />
     public bool ShouldHandle(ParameterInfo parameter)
-        => parameter.Parameter.Type!.ToString() == "bool";
+        => parameter.Parameter.Type!.ToString() == "string";
 
     /// <inheritdoc />
     public IError? GenerateSerializerPart(IndentedTextWriter textWriter, PacketInfo packet)
     {
         var parameter = packet.Parameters.Current;
-        if (parameter.Nullable)
-        {
-            textWriter.WriteLine($"if (obj.{parameter.Name} is null)");
-            textWriter.WriteLine("{");
-            textWriter.Indent++;
-            textWriter.WriteLine("builder.Append(\"-\");");
-            textWriter.Indent--;
-            textWriter.WriteLine("}");
-            textWriter.WriteLine("else");
-        }
-        textWriter.WriteLine("{");
-        textWriter.Indent++;
-        textWriter.WriteLine($"builder.Append(obj.{parameter.Name} ? \"1\" : \"0\");");
-        textWriter.Indent--;
-        textWriter.WriteLine("}");
+        textWriter.WriteLine($"builder.Append(obj.{parameter.Name} ?? \"-\");");
         return null;
     }
 
@@ -52,7 +38,7 @@ if ({parameter.GetErrorVariableName()} is not null)
 {{
     return Result<{packet.Name}?>.FromError({parameter.GetErrorVariableName()}, {parameter.GetResultVariableName()});
 }}
-{parameter.GetNullableType()} {parameter.GetNullableVariableName()} = {parameter.GetResultVariableName()}.Entity.Token == ""1"" ? true : ({parameter.GetResultVariableName()}.Entity.Token == ""-"" ? null : false);
+var {parameter.GetNullableVariableName()} = {parameter.GetResultVariableName()}.Entity.Token;
 ");
         return null;
     }
