+++ Core/NosSmooth.PacketSerializersGenerator/InlineTypeConverterGenerator.cs
@@ -8,7 +8,7 @@ using System.CodeDom.Compiler;
 using NosSmooth.PacketSerializersGenerator.Data;
 using NosSmooth.PacketSerializersGenerator.Errors;
 using NosSmooth.PacketSerializersGenerator.Extensions;
-using NosSmooth.PacketSerializersGenerator.TypeGenerators;
+using NosSmooth.PacketSerializersGenerator.InlineConverterGenerators;
 
 namespace NosSmooth.PacketSerializersGenerator;
